Move classes FillPattern, FilterRule and LinePattern from Elements to Views namespace
