+++ Revit_oM/Views/FilterRule.cs
@@ -5,11 +5,14 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.oM.Adapters.Revit.Elements
+namespace BH.oM.Revit.Views
 {
     public abstract class FilterRule : BHoMObject
     {
-        public virtual String ElementId { get; set; }
+        public virtual FilterRuleType RuleType { get; set; }
+        public virtual string ParameterName { get; set; }
+        public virtual object Value { get; set; }
+
     }
 
 }
